(version 1.2) add upload files sending email for forget password and when project status changed

+++ routes/projects.js
@@ -1,7 +1,10 @@
 const express = require('express');
-const { ProjectModel, validateProject, validateTask, validateEditProject, validateEditStatus } = require('../models/projectModel');
+const { ProjectModel, validateProject, validateTask, validateEditStatus } = require('../models/projectModel');
 const { authAdmin, auth } = require('../middlewares/auth');
 const { UserModel } = require('../models/userModel');
+const { upload } = require('../middlewares/uploadFiles');
+const fs = require('fs');
+const { sendEmail, informationUpdated } = require('../middlewares/sendEmail');
 const router = express.Router();
 
 router.get('/', auth, async (req, res) => {
@@ -81,6 +84,30 @@ router.post('/task/:projectID', authAdmin, async (req, res) => {
   }
 });
 
+router.put('/:projectID', authAdmin, async (req, res) => {
+  const validBody = validateProject(req.body);
+  if (validBody.error) {
+    return res.status(400).json(validBody.error.details);
+  }
+  try {
+    const projectID = req.params.projectID;
+    const project = await ProjectModel.findOne({ _id: projectID });
+    if (project.client_email !== req.body.client_email) {
+      const user = await UserModel.findOne({ email: req.body.client_email });
+      if (!user) return res.status(401).json({ err: 'client not found' });
+      project.client_id = user._id;
+      await project.save();
+    }
+    const update = await ProjectModel.updateOne({ _id: projectID }, req.body);
+
+    res.json(update);
+  } catch (err) {
+    console.log(err);
+    res.status(502).json({ err });
+  }
+});
+
+// TO DO change email
 router.put('/status/:projectID', authAdmin, async (req, res) => {
   const validBody = validateEditStatus(req.body);
   if (validBody.error) {
@@ -93,6 +120,8 @@ router.put('/status/:projectID', authAdmin, async (req, res) => {
     project.status.name = req.body.name;
     project.status.style = req.body.style;
     await project.save();
+
+    if (project.client_email) informationUpdated(project.client_email);
 
     res.status(201).json(project);
   } catch (err) {
@@ -142,9 +171,15 @@ router.put('/changeIsOpen/:projectID', authAdmin, async (req, res) => {
 router.delete('/project/:projectID', authAdmin, async (req, res) => {
   try {
     const projectID = req.params.projectID;
-    const project = await ProjectModel.deleteOne({ _id: projectID });
+    let project = await ProjectModel.findOne({ _id: projectID });
+    if (!project) return res.sendStatus(401);
+    if (project.file) {
+      const filePath = 'public/' + project.file;
+      await fs.promises.unlink(filePath);
+    }
+    const deleted = await ProjectModel.deleteOne({ _id: projectID });
 
-    res.status(200).json(project);
+    res.status(200).json(deleted);
   } catch (err) {
     console.log(err);
     res.status(502).json({ err });
@@ -164,6 +199,62 @@ router.delete('/task', authAdmin, async (req, res) => {
   } catch (err) {
     console.log(err);
     res.status(502).json({ err });
+  }
+});
+
+router.post('/uploadFile/:projectID', authAdmin, async (req, res) => {
+  try {
+    const projectID = req.params.projectID;
+    upload(req, res, async (err) => {
+      if (err || !req.file) {
+        console.log('err :' + err);
+        return res.status(400).json({ err: 'Error: jpeg|jpg|png|gif|pdf only!' });
+      } else {
+        let updateData = await ProjectModel.updateOne({ _id: projectID }, { file: 'files/' + req.file.filename });
+        res.json(updateData);
+      }
+    });
+  } catch (err) {
+    console.log(err);
+    res.status(502).json(err);
+  }
+});
+
+router.post('/editFile/:projectID', authAdmin, async (req, res) => {
+  try {
+    const projectID = req.params.projectID;
+    const project = await ProjectModel.findOne({ _id: projectID });
+    if (!project.file) return res.status(400).json({ err: 'No File found' });
+    const filePath = 'public/' + project.file;
+    await fs.promises.unlink(filePath);
+    upload(req, res, async (err) => {
+      if (err || !req.file) {
+        return res.status(400).json({ err: 'only File' });
+      } else {
+        let updateData = await ProjectModel.updateOne({ _id: projectID }, { file: 'file/' + req.file.filename });
+        res.json(updateData);
+      }
+    });
+  } catch (err) {
+    console.log(err);
+    res.status(502).json(err);
+  }
+});
+
+router.delete('/deleteFile/:projectID', authAdmin, async (req, res) => {
+  try {
+    const projectID = req.params.projectID;
+    let project = await ProjectModel.findOne({ _id: projectID });
+    if (!project) return res.sendStatus(401);
+    if (!project.file) return res.status(400).json({ err: 'No image found' });
+    const filePath = 'public/' + project.file;
+    await fs.promises.unlink(filePath);
+    user.file = null;
+    await project.save();
+    res.sendStatus(200);
+  } catch (err) {
+    console.log(err);
+    res.status(502).json(err);
   }
 });
